[Refactor] team 관련 소켓 event, handler 분리, event 이름 변경
team 관련한 소켓 handler 분리 리팩토링작업을 하였습니다.
