+++ backend/src/customeTypes/team.ts
@@ -1,0 +1,3 @@
+export interface UserType {
+	userId: number;
+}
